requirements.txt
Click==7.0
Flask==1.1.1
Flask-MySQLdb==0.2.0
GDAL==2.4.2
gunicorn==19.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
nose==1.3.7
numpy==1.16.4
protobuf==3.11.3
six==1.14.0
Werkzeug==1.0.0
